move functionality to app per spec
