+++ src/Button.module.css
@@ -1,0 +1,21 @@
+.base {
+  background-color: #eee;
+  border: 0;
+  border-radius: 4px;
+  cursor: pointer;
+  font-size: 1em;
+  padding: 0.4em 1em;
+  color: #000;
+  flex: 1 1;
+  min-width: max-content;
+}
+
+.base:disabled {
+  cursor: not-allowed;
+  opacity: 0.5;
+}
+
+.danger {
+  background-color: #e74c3c;
+  color: #fff;
+}
