+++ src/StopWatch.module.css
@@ -1,0 +1,3 @@
+.counter {
+  text-align: center;
+}
