--- src/StopWatchButton.module.css
@@ -1,8 +1,0 @@
-.base {
-  background-color: #fff;
-  border: 1px solid;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 1em;
-  padding: 0.5em 1em;
-}
