+++ src/StopWatchButtons.module.css
@@ -1,0 +1,6 @@
+.buttonRow {
+  display: flex;
+  gap: 0.5rem;
+  justify-content: space-between;
+  flex-wrap: wrap;
+}
