+++ webpack.common.js
@@ -30,13 +30,7 @@ module.exports = {
         }
     },{
         test: /\.css$/,
-        use: ['style-loader', {
-          loader: 'css-loader',
-          options: {
-            modules: true,
-            sourceMap: true,
-          }
-        }],
+        use: ['style-loader', 'css-loader'],
     },{
       test: /\.(png|jpe?g|gif|jp2|webp)$/,
       loader: 'file-loader',
